move the popup modal to another file to organize the code
